# gateaway-service/swagger.yaml
openapi: 3.0.0
info:
  title: AI-Inventory-System API
  description: API Gateway documentation for the Microservices-based Inventory System.
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local API Gateway
paths:
...
  /api/sales:
    get:
      summary: Ambil Riwayat Penjualan
      tags: [Sales]
      security:
        - bearerAuth: []
      responses:
        200:
          description: List of sales transactions
  /api/sales/transaction:
    post:
      summary: Buat Transaksi Penjualan Baru
      tags: [Sales]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
                quantity_sold:
                  type: integer
                sales_person:
                  type: string
      responses:
        201:
          description: Sale recorded

  # --- STOCK SERVICE ---
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT